+++ src/common/Controller.js
@@ -5,5 +5,7 @@ export default class Controller {
     this.model = model;
     this.view = view;
     this.$router = Router.instance;
+    this.init();
   }
+  init() {}
 }
