+++ src/common/View.js
@@ -1,8 +1,11 @@
+import Router from '../router';
+
 export default class View {
   constructor(target) {
     this.$newEl = target.cloneNode(true);
     this.$newEl.innerHTML = this.getTemplate();
     target.replaceWith(this.$newEl);
+    this.$router = Router.instance;
   }
 
   addComponent(className, component) {
