--- src/view/layouts/Header/headerViewTemplate.js
@@ -1,17 +1,0 @@
-export default () => `
-  <nav>
-    <ul>
-      <div class="router_section">
-        <li>
-          <a href="/question" data-route="/question">질문 받기</a>
-        </li>
-        <li>
-          <a href="/manage" data-route="/manage">질문 관리하기</a>
-        </li>
-      </div>
-      <li>
-        <a href="https://github.com/pozafly" target="_blank" class="github_link">Github</a>
-      </li>
-    </ul>
-  </nav>
-`;
