--- src/view/layouts/layoutViewTemplate.js
@@ -1,5 +1,0 @@
-export default () => `
-  <header></header>
-  <main></main>
-  <footer></footer>
-`;
