--- src/view/manage/ManageViewTemplate.js
@@ -1,3 +1,0 @@
-export default () => `
-  <div>Manage Page</div>
-`;
